test: add testing
