+++ src/components/AddProcessButton/Dialog/index.ts
@@ -1,0 +1,1 @@
+export { AddProcessButtonDialog } from './Dialog'
